requirements.txt
streamlit==1.28.1
streamlit-webrtc==0.47.1
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
gtts==2.4.0
speechrecognition==3.10.0
transformers==4.35.2
torch==2.1.0
torchvision==0.16.0
googletrans==4.0.0rc1
geopy==2.4.0
plotly==5.17.0
streamlit-option-menu==0.3.6
audio-recorder-streamlit==0.0.8
pydub==0.25.1